translated to using fragments
