+++ Asay.mobile/app/src/main/res/layout/activity_main.xml
@@ -4,7 +4,10 @@
     android:layout_height="match_parent"
     android:orientation="vertical" >
 
-    <ListView
+    <FrameLayout
+        android:id="@+id/billListFrame"
         android:layout_width="match_parent"
-        android:layout_height="match_parent" />
+        android:layout_height="match_parent">
+
+    </FrameLayout>
 </LinearLayout>
